requirements.txt
coverage==3.7.1
utifulsoup4==4.3.2
Django==1.8.2
django-tagging==0.4
django-webtest==1.7.8
selenium==2.46.0
six==1.9.0
waitress==0.8.9
WebOb==1.4.1
WebTest==2.0.18

